Add MP list spider to scrape list of MPs
A new spider "list_of_mps" is added to scrape list of all MPs. For debugging
another spider is added which scrapes the data and saves it to a local file. We
can use this file to avoid repeated GET requests while debugging and this spider
can be deleted at the end. At the moment the list of all MPs are only printed on
to console.

Signed-off-by: Arun Siluvery <arunk.siluvery@gmail.com>